# Conflicts:
#	src/InjectImageIntrinsics.cpp


Former-commit-id: 132e3046967131c04587cb4a9eab8ca5b89cc868

+++ src/runtime/HalideRuntimeOpenCL.h
@@ -68,17 +68,16 @@ extern const char *halide_opencl_get_device_type(void *user_context);
  * routine is called. This call can fail due to running out of memory
  * or being passed an invalid device pointer. The device and host
  * dirty bits are left unmodified. */
-extern int halide_opencl_wrap_cl_mem(void *user_context, struct halide_buffer_t *buf, uintptr_t device_ptr);
+extern int halide_opencl_wrap_cl_mem(void *user_context, struct halide_buffer_t *buf, uint64_t device_ptr);
 
-/** Disconnect a halide_buffer_t from the memory it was previously wrapped
- * around. Should only be called for a halide_buffer_t that
+/** Disconnect a halide_buffer_t from the memory it was previously
+ * wrapped around. Should only be called for a halide_buffer_t that
  * halide_opencl_wrap_device_ptr was previously called on. Frees any
- * storage associated with the binding of the halide_buffer_t and the device
- * pointer, but does not free the cl_mem. The previously wrapped
- * cl_mem is returned. The dev field of the halide_buffer_t will be NULL on
- * return.
+ * storage associated with the binding of the halide_buffer_t and the
+ * device pointer, but does not free the cl_mem. The dev field of the
+ * halide_buffer_t will be NULL on return.
  */
-extern uintptr_t halide_opencl_detach_cl_mem(void *user_context, struct halide_buffer_t *buf);
+extern int halide_opencl_detach_cl_mem(void *user_context, struct halide_buffer_t *buf);
 
 /** Return the underlying cl_mem for a halide_buffer_t. This buffer must be
  *  valid on an OpenCL device, or not have any associated device
